PANDA.jl Phylogenetic ANalyses of DiversificAtion

Implements macroevolutionary analyses on phylogenetic trees.

Installation

You can install the package by typing

julia> using Pkg
julia> Pkg.add("PANDA")

Once installed, it can be loaded to your environment with

julia> using PANDA

ClaDS

The ClaDS module implements the inference of ClaDS parameters on a phylogeny using data augmentation.

Module = [PANDA.ClaDS]

PANDA.ClaDS.load_tree — Function

load_tree(file::String)

Loads a tree from a file. Supported formats are '.nex' and '.tre'.

Arguments

file::String:path to the file.

Missing docstring.

Missing docstring for PANDA.ClaDS.infer_ClaDS. Check Documenter's build log for details.

PANDA.ClaDS.plot_ClaDS — Function

plot_ClaDS(tree::Tree ; ln=true, show_labels = false, options="")

Plots a tree with branches colored according to their rates

Arguments

tree::Tree: the phylogeny to be plotted.

Keyword arguments

ln::Bool: Should rates be plotted on a log scale? Default to true.

show_labels::Bool: Should tip labels be printed? Default to false.

options::String: Additonal options for the ploting function.

plot_ClaDS(tree::Tree, rates::Array{Number,1} ; ln=true, show_labels = false, options="")

Additional arguments

rates::Array{Number,1}: the speciation rates.

Missing docstring.

Missing docstring for PANDA.ClaDS.sim_ClaDS2_ntips. Check Documenter's build log for details.