PANDA.ClaDS.infer_ClaDS — Function

infer_ClaDS(tree::Tree, n_reccord::Int64; ...)

Infer ClaDS parameters on a tree

Arguments

tree::Tree: the phylogeny on which to perform the inference.

n_reccord::Int64: number of iterations between two computations of the gelman statistics. Default to 1_000.

Keyword arguments

former_run::CladsOutput: the result of a run of infer_ClaDS on the tree. The new mcmc chains will be added to these ones.

f::Float64: The sampling probability. Default to 1..

thin::Int64: The thinning parameter. Default to 1..

burn::Float64: The proportion of the mcmc that will be discarded befor computing the gelman statistics and point estiamtes for the parameters. Default to 0.25.

n_trees::Int64: Number of samples from the posterior distribution of complete phylogenies to be outputed. Default to 10.

goal_gelman::Float64: . Default to 1.05.