PANDA.jl Phylogenetic ANalyses of DiversificAtion

Implements macroevolutionary analyses on phylogenetic trees.

Installation

You can install the package by typing

julia> using Pkg
julia> Pkg.add("PANDA")

Once installed, it can be loaded to your environment with

julia> using PANDA

ClaDS

The ClaDS module implements the inference of ClaDS parameters on a phylogeny using data augmentation.

PANDA.ClaDS.Tree — Type

A phylogeny object for the module ClaDS. It is represented as a branch with a given length and optional attributes, and its daughter trees.

offsprings::Array{Tree,1}: the two daughter trees.

branch_length::Float64: the length of the branch

attributes::Array{T,1} where {T<:Number}: soem attributes of the branch. Used to store the speciation rates.

n_nodes::Int64: the number of nodes in the tree (internal nodes + tips)

extant::Bool: does the tree have any extant species?

label::String: if the tree is a tip (ie n_nodes == 1), contains the name of the corresponding species.

PANDA.ClaDS.extract_branch_lengths — Method

extract_branch_lengths(tree::Tree)

extract the branch lengths from a tree object, in the preorder order

PANDA.ClaDS.get_node_depth — Method

get_node_depth(tree::Tree, edge_id::Int64)

extract the node depth from a tree object (distance to the present), in the preorder order

PANDA.ClaDS.infer_ClaDS — Function

infer_ClaDS(tree::Tree, n_reccord::Int64; ...)

Infer ClaDS parameters on a tree

Arguments

tree::Tree: the phylogeny on which to perform the inference.

n_reccord::Int64: number of iterations between two computations of the gelman statistics. Default to 1_000.

Keyword arguments

former_run::CladsOutput: the result of a run of infer_ClaDS on the tree. The new mcmc chains will be added to these ones.

f::Float64: The sampling probability. Default to 1..

thin::Int64: The thinning parameter. Default to 1..

burn::Float64: The proportion of the mcmc that will be discarded befor computing the gelman statistics and point estiamtes for the parameters. Default to 0.25.

n_trees::Int64: Number of samples from the posterior distribution of complete phylogenies to be outputed. Default to 10.

goal_gelman::Float64: . Default to 1.05.

PANDA.ClaDS.n_tips — Method

n_tips(tree::Tree)

Get the number of tip in a tree.

Arguments

tree::Tree: a Treeobject.

PANDA.ClaDS.plot_ClaDS — Method

plot_ClaDS(tree::Tree ; ln=true, show_labels = false, options="")

Plots a tree with branches colored according to their rates

Arguments

tree::Tree: the phylogeny to be plotted.

Keyword arguments

ln::Bool: Should rates be plotted on a log scale? Default to true.

show_labels::Bool: Should tip labels be printed? Default to false.

options::String: Additonal options for the ploting function.

plot_ClaDS(tree::Tree, rates::Array{Number,1}  ; ln=true, show_labels = false, options="")

Arguments

rates::Array{Number,1}: the speciation rates.

PANDA.ClaDS.sim_ClaDS2_ntips — Method

sim_ClaDS2_ntips(n::Int64,σ::Float64,α::Float64,ε::Float64,λ0::Float64 ; ...)

Simulate a tree from the ClaDS model conditionned on the number of tips at present

Arguments

n::Int64: the goal number of tips at present.

σ::Float64: the stochasticity parameter.

α::Float64: the trend parameter.

ε::Float64: turnover rate (extinction / speciation).

λ0::Float64: initial speciation rate.

Keyword arguments

prune_extinct::Bool: Should extinct lineages be removed from the output tree? Default to false.

PANDA.ClaDS.tip_labels — Method

tip_labels(tree::Tree)

Extract the tip labels of a tree.

Arguments

tree::Tree: a Treeobject.